[move/#229] EmptyTextToolbar 파일 이동
