--- core/util/src/main/java/com/napzak/market/util/android/ContentUriRequestBody.kt
@@ -1,2 +1,0 @@
-package com.napzak.market.util.android
-
