+++ feature/registration/build.gradle.kts
@@ -14,6 +14,7 @@ dependencies {
     implementation(projects.core.designsystem)
     implementation(projects.core.common)
     implementation(projects.core.util)
+    implementation(projects.core.uiUtil)
 
     implementation(projects.domain.registration)
     implementation(projects.domain.presignedUrl)
